pyproject.toml
[project]
name = "skill-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for managing Claude skills"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
skill-mcp-server = "skill_mcp.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/skill-mcp"
Documentation = "https://github.com/yourusername/skill-mcp#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/skill_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

